academic_transfer/requirements.txt
Flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
python-docx==1.1.2